Rename projection to consumers
