+++ packages/server/events-processing-platform/consumers/email_validation/email_event_handlers.go
