+++ packages/server/events-processing-platform/consumers/graph/email_event_handlers.go
