+++ packages/server/events-processing-platform/consumers/worker.go
@@ -1,4 +1,4 @@
-package projection
+package consumers
 
 import (
 	"github.com/EventStore/EventStore-Client-Go/v3/esdb"
